refactor: Re-organazing pieces of code related to typechecking
